refactored SecurityConfig into separate classes & implemented CustomUserDetails
