+++ src/main/java/org/orange/oie/internship2025/assetmanagementsystem/entity/Role.java
@@ -1,8 +1,7 @@
 package org.orange.oie.internship2025.assetmanagementsystem.entity;
+
 import jakarta.persistence.*;
 import lombok.*;
-
-import java.util.List;
 
 @Getter
 @Setter
