+++ src/main/java/org/orange/oie/internship2025/assetmanagementsystem/entity/Type.java
@@ -1,7 +1,7 @@
 package org.orange.oie.internship2025.assetmanagementsystem.entity;
+
 import jakarta.persistence.*;
 import lombok.*;
-
 
 @Getter
 @Setter
