+++ src/main/java/org/orange/oie/internship2025/assetmanagementsystem/mapper/UserMapper.java
@@ -2,14 +2,13 @@ package org.orange.oie.internship2025.assetmanagementsystem.mapper;
 
 import org.orange.oie.internship2025.assetmanagementsystem.dto.UserDTO;
 import org.orange.oie.internship2025.assetmanagementsystem.entity.User;
-import org.springframework.context.annotation.Profile;
 import org.springframework.stereotype.Component;
 
 @Component
 
 public class UserMapper {
 
-    public static UserDTO toDto(User user) {
+    public UserDTO toDto(User user) {
         if (user == null) {
             return null;
         }
@@ -20,8 +19,6 @@ public class UserMapper {
         userDTO.setUsername(user.getUsername());
         userDTO.setRole(user.getRole());
         userDTO.setDepartmentName(user.getDepartment().getDepartmentName());
-
-
 
         return userDTO;
     }
