+++ src/main/java/org/orange/oie/internship2025/assetmanagementsystem/security/CustomUserDetails.java
@@ -1,0 +1,57 @@
+package org.orange.oie.internship2025.assetmanagementsystem.security;
+
+import org.orange.oie.internship2025.assetmanagementsystem.entity.User;
+import org.springframework.security.core.GrantedAuthority;
+import org.springframework.security.core.authority.SimpleGrantedAuthority;
+import org.springframework.security.core.userdetails.UserDetails;
+
+import java.util.Collection;
+import java.util.List;
+
+public class CustomUserDetails implements UserDetails {
+
+    private final User user;
+
+    public CustomUserDetails(User user) {
+        this.user = user;
+    }
+
+    @Override
+    public Collection<? extends GrantedAuthority> getAuthorities() {
+        return List.of(new SimpleGrantedAuthority(user.getRole().getRoleType()));
+    }
+
+    @Override
+    public String getPassword() {
+        return user.getPassword();
+    }
+
+    @Override
+    public String getUsername() {
+        return user.getEmail(); // use email as username
+    }
+
+    @Override
+    public boolean isAccountNonExpired() {
+        return true; // customize if you track expiry
+    }
+
+    @Override
+    public boolean isAccountNonLocked() {
+        return true; // customize if you track lock status
+    }
+
+    @Override
+    public boolean isCredentialsNonExpired() {
+        return true; // customize if needed
+    }
+
+    @Override
+    public boolean isEnabled() {
+        return true; // customize if needed
+    }
+
+    public User getUser() {
+        return user; // so you can access full entity if needed
+    }
+}
