+++ src/main/java/org/orange/oie/internship2025/assetmanagementsystem/security/CustomUserDetailsService.java
@@ -1,0 +1,32 @@
+package org.orange.oie.internship2025.assetmanagementsystem.security;
+
+import org.orange.oie.internship2025.assetmanagementsystem.entity.User;
+import org.orange.oie.internship2025.assetmanagementsystem.repository.UserRepository;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.security.core.userdetails.UserDetails;
+import org.springframework.security.core.userdetails.UserDetailsService;
+import org.springframework.security.core.userdetails.UsernameNotFoundException;
+import org.springframework.stereotype.Service;
+
+@Service
+public class CustomUserDetailsService implements UserDetailsService {
+
+    @Autowired
+    private UserRepository userRepository;
+
+    @Override
+    public UserDetails loadUserByUsername(String username) throws UsernameNotFoundException {
+
+        if (username == null || !username.endsWith("@orange.com")) {
+            throw new UsernameNotFoundException("Invalid domain for user: " + username);
+        }
+
+        User user = userRepository.findByEmail(username);
+        if (user == null) {
+            throw new UsernameNotFoundException("User not found: " + username);
+        }
+
+        return new CustomUserDetails(user);
+    };
+
+}
