+++ src/main/java/org/orange/oie/internship2025/assetmanagementsystem/security/PasswordConfig.java
@@ -1,0 +1,16 @@
+
+package org.orange.oie.internship2025.assetmanagementsystem.security;
+
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
+import org.springframework.security.crypto.password.PasswordEncoder;
+
+@Configuration
+public class PasswordConfig {
+
+    @Bean
+    public PasswordEncoder passwordEncoder() {
+        return new BCryptPasswordEncoder();
+    }
+}
