+++ src/main/java/org/orange/oie/internship2025/assetmanagementsystem/service/AuthService.java
@@ -4,28 +4,23 @@ import org.orange.oie.internship2025.assetmanagementsystem.dto.UserDTO;
 import org.orange.oie.internship2025.assetmanagementsystem.entity.User;
 import org.orange.oie.internship2025.assetmanagementsystem.mapper.UserMapper;
 import org.orange.oie.internship2025.assetmanagementsystem.repository.UserRepository;
-import org.springframework.security.authentication.BadCredentialsException;
+import org.orange.oie.internship2025.assetmanagementsystem.security.CustomUserDetails;
 import org.springframework.security.core.Authentication;
-import org.springframework.security.core.userdetails.UsernameNotFoundException;
 import org.springframework.stereotype.Service;
 
 @Service
 public class AuthService {
 
-    private final UserRepository userRepository;
     private final UserMapper userMapper;
 
-    public AuthService(UserRepository userRepository, UserMapper userMapper) {
-        this.userRepository = userRepository;
+    public AuthService(UserMapper userMapper) {
+
         this.userMapper = userMapper;
     }
 
     public final UserDTO authenticateUser(Authentication authentication) {
-        String email = authentication.getName();
-
-        User user = userRepository.findByEmail(email);
-
-        return UserMapper.toDto(user);
+        CustomUserDetails userdetails = (CustomUserDetails) authentication.getPrincipal();
+        return userMapper.toDto(userdetails.getUser());
     }
 
 }
